+++ lib/active_storage_validations/dimension_validator.rb
@@ -4,25 +4,30 @@ require_relative 'metadata.rb'
 
 module ActiveStorageValidations
   class DimensionValidator < ActiveModel::EachValidator # :nodoc
+    include OptionProcUnfolding
+
     AVAILABLE_CHECKS = %i[width height min max].freeze
 
-    def initialize(options)
+    def process_options(record)
+      flat_options = unfold_procs(record, self.options, AVAILABLE_CHECKS)
+
       [:width, :height].each do |length|
-        if options[length] and options[length].is_a?(Hash)
-          if range = options[length][:in]
+        if flat_options[length] and flat_options[length].is_a?(Hash)
+          if (range = flat_options[length][:in])
             raise ArgumentError, ":in must be a Range" unless range.is_a?(Range)
-            options[length][:min], options[length][:max] = range.min, range.max
+            flat_options[length][:min], flat_options[length][:max] = range.min, range.max
           end
         end
       end
       [:min, :max].each do |dim|
-        if range = options[dim]
+        if (range = flat_options[dim])
           raise ArgumentError, ":#{dim} must be a Range (width..height)" unless range.is_a?(Range)
-          options[:width] = { dim => range.first }
-          options[:height] = { dim => range.last }
+          flat_options[:width] = { dim => range.first }
+          flat_options[:height] = { dim => range.last }
         end
       end
-      super
+
+      flat_options
     end
 
 
@@ -64,26 +69,27 @@ module ActiveStorageValidations
 
 
     def is_valid?(record, attribute, file_metadata)
+      flat_options = process_options(record)
       # Validation fails unless file metadata contains valid width and height.
       if file_metadata[:width].to_i <= 0 || file_metadata[:height].to_i <= 0
-        add_error(record, attribute, options[:message].presence || :image_metadata_missing)
+        add_error(record, attribute, :image_metadata_missing)
         return false
       end
 
       # Validation based on checks :min and :max (:min, :max has higher priority to :width, :height).
-      if options[:min] || options[:max]
-        if options[:min] && (
-          (options[:width][:min] && file_metadata[:width] < options[:width][:min]) ||
-          (options[:height][:min] && file_metadata[:height] < options[:height][:min])
+      if flat_options[:min] || flat_options[:max]
+        if flat_options[:min] && (
+          (flat_options[:width][:min] && file_metadata[:width] < flat_options[:width][:min]) ||
+          (flat_options[:height][:min] && file_metadata[:height] < flat_options[:height][:min])
           )
-          add_error(record, attribute, options[:message].presence || :"dimension_min_inclusion", width: options[:width][:min], height: options[:height][:min])
+          add_error(record, attribute, :dimension_min_inclusion, width: flat_options[:width][:min], height: flat_options[:height][:min])
           return false
         end
-        if options[:max] && (
-          (options[:width][:max] && file_metadata[:width] > options[:width][:max]) ||
-          (options[:height][:max] && file_metadata[:height] > options[:height][:max])
+        if flat_options[:max] && (
+          (flat_options[:width][:max] && file_metadata[:width] > flat_options[:width][:max]) ||
+          (flat_options[:height][:max] && file_metadata[:height] > flat_options[:height][:max])
           )
-          add_error(record, attribute, options[:message].presence || :"dimension_max_inclusion", width: options[:width][:max], height: options[:height][:max])
+          add_error(record, attribute, :dimension_max_inclusion, width: flat_options[:width][:max], height: flat_options[:height][:max])
           return false
         end
 
@@ -91,24 +97,24 @@ module ActiveStorageValidations
       else
         width_or_height_invalid = false
         [:width, :height].each do |length|
-          next unless options[length]
-          if options[length].is_a?(Hash)
-            if options[length][:in] && (file_metadata[length] < options[length][:min] || file_metadata[length] > options[length][:max])
-              add_error(record, attribute, options[:message].presence || :"dimension_#{length}_inclusion", min: options[length][:min], max: options[length][:max])
+          next unless flat_options[length]
+          if flat_options[length].is_a?(Hash)
+            if flat_options[length][:in] && (file_metadata[length] < flat_options[length][:min] || file_metadata[length] > flat_options[length][:max])
+              add_error(record, attribute, :"dimension_#{length}_inclusion", min: flat_options[length][:min], max: flat_options[length][:max])
               width_or_height_invalid = true
             else
-              if options[length][:min] && file_metadata[length] < options[length][:min]
-                add_error(record, attribute, options[:message].presence || :"dimension_#{length}_greater_than_or_equal_to", length: options[length][:min])
+              if flat_options[length][:min] && file_metadata[length] < flat_options[length][:min]
+                add_error(record, attribute, :"dimension_#{length}_greater_than_or_equal_to", length: flat_options[length][:min])
                 width_or_height_invalid = true
               end
-              if options[length][:max] && file_metadata[length] > options[length][:max]
-                add_error(record, attribute, options[:message].presence || :"dimension_#{length}_less_than_or_equal_to", length: options[length][:max])
+              if flat_options[length][:max] && file_metadata[length] > flat_options[length][:max]
+                add_error(record, attribute, :"dimension_#{length}_less_than_or_equal_to", length: flat_options[length][:max])
                 width_or_height_invalid = true
               end
             end
           else
-            if file_metadata[length] != options[length]
-              add_error(record, attribute, options[:message].presence || :"dimension_#{length}_equal_to", length: options[length])
+            if file_metadata[length] != flat_options[length]
+              add_error(record, attribute, :"dimension_#{length}_equal_to", length: flat_options[length])
               width_or_height_invalid = true
             end
           end
@@ -120,10 +126,10 @@ module ActiveStorageValidations
       true # valid file
     end
 
-    def add_error(record, attribute, type, **attrs)
-      key = options[:message].presence || type
-      return if record.errors.added?(attribute, key)
-      record.errors.add(attribute, key, **attrs)
+    def add_error(record, attribute, default_message, **attrs)
+      message = options[:message].presence || default_message
+      return if record.errors.added?(attribute, message)
+      record.errors.add(attribute, message, **attrs)
     end
 
   end