+++ test/dummy/app/models/user.rb
@@ -13,6 +13,10 @@ class User < ApplicationRecord
   has_many_attached :photos
   has_one_attached :image_regex
   has_one_attached :conditional_image
+  has_one_attached :conditional_image_2
+  has_one_attached :proc_avatar
+  has_many_attached :proc_photos
+  has_one_attached :proc_image_regex
 
   validates :name, presence: true
 
@@ -20,4 +24,8 @@ class User < ApplicationRecord
   validates :photos, attached: true, content_type: ['image/png', 'image/jpg', /\A.*\/pdf\z/]
   validates :image_regex, content_type: /\Aimage\/.*\z/
   validates :conditional_image, attached: true, if: -> { name == 'Foo' }
+  validates :conditional_image_2, attached: true, content_type: -> (record) {[/\Aimage\/.*\z/]}, size: { less_than: 10.megabytes }, if: -> { name == 'Peter Griffin' }
+  validates :proc_avatar, attached: { message: "must not be blank" }, content_type: -> (record) {:png}
+  validates :proc_photos, attached: true, content_type: -> (record) {['image/png', 'image/jpg', /\A.*\/pdf\z/]}
+  validates :proc_image_regex, content_type: -> (record) {/\Aimage\/.*\z/}
 end