+++ test/matchers/content_type_validator_matcher_test.rb
@@ -4,10 +4,25 @@ require 'test_helper'
 require 'active_storage_validations/matchers'
 
 class ActiveStorageValidations::Matchers::ContentTypeValidatorMatcher::Test < ActiveSupport::TestCase
+  test 'positive match on both allowing and rejecting' do
+    matcher = ActiveStorageValidations::Matchers::ContentTypeValidatorMatcher.new(:avatar)
+    matcher.allowing('image/png')
+    matcher.rejecting('image/jpg')
+
+    assert matcher.matches?(User)
+  end
+
+  test 'negative match on both allowing and rejecting' do
+    matcher = ActiveStorageValidations::Matchers::ContentTypeValidatorMatcher.new(:avatar)
+    matcher.rejecting('image/png')
+    matcher.allowing('image/jpg')
+
+    refute matcher.matches?(User)
+  end
+
   test 'positive match when providing class' do
     matcher = ActiveStorageValidations::Matchers::ContentTypeValidatorMatcher.new(:avatar)
     matcher.allowing('image/png')
-    matcher.rejecting('application/pdf')
     assert matcher.matches?(User)
   end
 
@@ -26,7 +41,6 @@ class ActiveStorageValidations::Matchers::ContentTypeValidatorMatcher::Test < Ac
   test 'positive match when providing instance' do
     matcher = ActiveStorageValidations::Matchers::ContentTypeValidatorMatcher.new(:avatar)
     matcher.allowing('image/png')
-    matcher.rejecting('application/pdf')
     assert matcher.matches?(User.new)
   end
 
@@ -40,5 +54,23 @@ class ActiveStorageValidations::Matchers::ContentTypeValidatorMatcher::Test < Ac
     matcher = ActiveStorageValidations::Matchers::ContentTypeValidatorMatcher.new(:non_existing)
     matcher.allowing('image/png')
     refute matcher.matches?(User.new)
+  end
+
+  test 'positive match for rejecting' do
+    matcher = ActiveStorageValidations::Matchers::ContentTypeValidatorMatcher.new(:avatar)
+    matcher.rejecting('image/jpeg')
+    assert matcher.matches?(User)
+  end
+
+  test 'negative match for rejecting' do
+    matcher = ActiveStorageValidations::Matchers::ContentTypeValidatorMatcher.new(:avatar)
+    matcher.rejecting('image/png')
+    refute matcher.matches?(User)
+  end
+
+  test 'positive match on subset of accepted content types' do
+    matcher = ActiveStorageValidations::Matchers::ContentTypeValidatorMatcher.new(:photos)
+    matcher.allowing('image/png')
+    assert matcher.matches?(User)
   end
 end